% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_survival.R
\name{basic_survival}
\alias{basic_survival}
\title{basic_survival}
\usage{
basic_survival(indiv_data, mps, lvps)
}
\arguments{
\item{indiv_data}{current indiv_data dataframe}

\item{mps}{Mean survival probability per demographic timestep (between 0 and 1)}

\item{lvps}{Variance in survival probability on a logit scale}
}
\value{
The new indiv_data dataframe with newly calculated survival probabilities added
}
\description{
Generates survival probabilities for all individuals in the population based on draws from a Gaussian distribution on a logit scale.
}
\details{
Generates survival probabilities for all individuals in the population based on draws from a Gaussian distribution on a logit scale.
}
