% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_and_obs2.R
\name{cap_and_obs2}
\alias{cap_and_obs2}
\title{cap_and_obs2}
\usage{
cap_and_obs2(
  samp_wind,
  gbi = gbi,
  pcg = 0.5,
  pmg = 0.5,
  pmi = 0.9,
  pci = 0.9,
  obs_seq = 1,
  cap_seq = 1,
  pre_cap = NULL
)
}
\arguments{
\item{samp_wind}{vector recording which groups correspond to which sampling windows/behavioural timesteps.}

\item{gbi}{the true group-by-individual matrix}

\item{pcg}{probability of observing or capturing a group in a sampling window with captures}

\item{pmg}{probability of observing or observing a group in a sampling window with observations}

\item{pmi}{the probability of an individual being observed in a sampled group}

\item{pci}{the probability of an individual being captured in a sampled group}

\item{obs_seq}{is an ordered vector of behavioural timesteps that observations take place in, e.g. c(2,4) would indicate observations in the second and fourth timesteps}

\item{cap_seq}{is an ordered vector of behavioural timesteps that captures take place in, e.g. c(2,4) would indicate captures in the second and fourth timesteps}

\item{pre_cap}{vector indicating whether any individuals have been previously captured (and so are observable). Defaults to NULL.}
}
\value{
A list with 4 elements: A) the full GBI but for only captured groups; B) a vector indicating which groups were captured ; C) the full GBI for observed and captured groups combined; D) a vector indicating which groups were observed
}
\description{
Generates observations that can be used to generate CMR data and observed social networks.
}
\details{
Note that observations do not happen in the same sampling windows as captures even if defined to do so. Only previously captured individuals are observable. See notes in comments for future updates. I would also like to allow some observation error in another version of this function.
}
