% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_dat_gen.R
\name{cap_dat_gen}
\alias{cap_dat_gen}
\title{cap_dat_gen}
\usage{
cap_dat_gen(CG, SW_store, full_indiv_data, inds_alive, bs)
}
\arguments{
\item{CG}{a list of the observed GBIs (element 3 from the output of cap_and_obs) for all demographic timesteps.}

\item{SW_store}{a list of all sampling windows (element 2 from the output of interaction_generation_simul)}

\item{full_indiv_data}{the full_indiv_data dataframe containing information on the ID, group and spatial location of all individuals recorded in the population}

\item{inds_alive}{a list of individuals recorded alive at each demographic timestep.}

\item{bs}{number of behavioural timesteps per demographic timestep.}
}
\value{
A list of 2 elements: A) Capture history summarised by demographic timesteps; B) Capture history for all behavioural timesteps
}
\description{
Function to generate typical CMR datasets from the output of cap_and_obs.
}
\details{
used to create typical CMR capture history datasets from the output of cap_and_obs for use in hidden Markov modelling.
}
