% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_survival.R
\name{covariates_survival}
\alias{covariates_survival}
\title{covariates_survival}
\usage{
covariates_survival(
  indiv_data,
  indiv_info,
  network,
  group_means = NULL,
  ext_vars = NULL,
  ext_effs,
  scale_ext = TRUE,
  net_vars = NULL,
  net_effs,
  net_packages,
  scale_net = TRUE,
  net_cov = FALSE,
  max_cor = 0.8,
  covMat_check = FALSE,
  mps,
  lvps
)
}
\arguments{
\item{indiv_data}{current indiv_data dataframe}

\item{indiv_info}{current indiv_info}

\item{network}{adjacency matrix for the network used to calculate survival probabilities}

\item{group_means}{social group (as defined in population generation) effects on survival probabilies. Provided on logit scale.}

\item{ext_vars}{vector extrinsic variables (non-network) used to calculate survival probability. Names of columns in indiv_info dataframe. Defaults to NULL.}

\item{ext_effs}{a list of effect sizes for each extrinsic variable provided. Single value for covariates. Value for each level for factors. Defaults to NULL}

\item{scale_ext}{whether to scale extrinsic variables. Defaults to TRUE.}

\item{net_vars}{vector of network measures that affect survival. Should be named as per the name of the function used in R (for a given package provided in net_packages=). Defaults to NULL}

\item{net_effs}{list of effect sizes used to calculate effect of network measures on survival}

\item{net_packages}{a vector of R packages of the same length of the net_vars providing the R package used to calculate each network measure. See examples}

\item{scale_net}{whether to scale network measures. Defaults to TRUE.}

\item{net_cov}{whether there should be network covariance in survival probability. Defaults to FALSE.}

\item{max_cor}{maximum correlation (positive or negative) for network covariance in survival probabilities}

\item{covMat_check}{whether to calculate a network regression to demonstrate that the adjusted covariance matrix (adjusted to be positive definite) is strongly correlated with the input matrix. Defaults to FALSE.}

\item{mps}{overall mean population survival probability}

\item{lvps}{variance in survival probability on logit scale}
}
\value{
The new indiv_data dataframe with newly calculated survival probabilities added
}
\description{
Generates survival probabilities for all individuals in the population based on their individual traits and position in the population social network.
}
\details{
This function simulates survival probabilities for all individuals in the population based on the covariates provided. Note that there is a huge amount of flexibility in this function, especially the network measures part of it but that its impossible to anticipate all possible demands with some packages/functions requiring varied inputs. The main ones are already covered but any issues with this can be reported as bugs and the package can be updated. Note also that the covariance matrix must be positive definite. We use the R function nearPD from the Matrix package to find the closest matrix to the provided network that is positive definite. This approximation normally works well but there is capability included to check how well correlated it is with the input network.
}
