% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_calc.R
\name{group_calc}
\alias{group_calc}
\title{group_calc}
\usage{
group_calc(ni, mgs)
}
\arguments{
\item{ni}{number of individuals}

\item{mgs}{mean group size}
}
\value{
vector of group sizes
}
\description{
internal function to generate group sizes
}
