% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_int.R
\name{ib_int}
\alias{ib_int}
\title{ib_int}
\usage{
ib_int(a, b)
}
\arguments{
\item{a}{an non-integer number}

\item{b}{a second non-integer number (higher)}
}
\value{
a vector of the integers between them
}
\description{
internal function to find integers between two non-integer number
}
\details{
internal function to find integers between two non-integer number
}
