% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_info_add.R
\name{indiv_info_add}
\alias{indiv_info_add}
\title{indiv_info_add}
\usage{
indiv_info_add(indiv_info, ii_tag, indiv_data)
}
\arguments{
\item{indiv_info}{existing indiv_info dataframe.}

\item{ii_tag}{information on variables already included in the dataframe.}

\item{indiv_data}{dataframe containing data on individual IDs, spatial locations and social group IDs}
}
\value{
A nested list. The first element of the list contains the dataframe of individual data. Each subsequent element contains a list of information about each explanatory variable added.
}
\description{
Adds trait data for individuals at a demographic timestep
}
\details{
Adds trait data (e.g. sex, body size etc.) for newly recruited individuals.
}
