% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_info_gen.R
\name{indiv_info_gen}
\alias{indiv_info_gen}
\title{indiv_info_gen}
\usage{
indiv_info_gen(
  indiv_info,
  ii_tag = NULL,
  indiv_data,
  trait,
  trait_type = c("cov", "fac"),
  x_dist = c("norm"),
  level_names = NULL,
  prob_levels = NULL
)
}
\arguments{
\item{indiv_info}{existing indiv_info dataframe. Minimum input is a single-column dataframe of individual IDs}

\item{ii_tag}{information on variables already included in the dataframe if present. Defaults to NULL}

\item{indiv_data}{dataframe from population_generation_basic() containing data on individual IDs, spatial locations and social group IDs}

\item{trait}{name of trait to be added}

\item{trait_type}{either a covariate (“cov”) or factor (“fac”)}

\item{x_dist}{the distribution of a covariate. Either uniform (“unif”), normal (“norm”) or log-normal (“lnorm”).}

\item{level_names}{the names of factor levels for categorical variables. Defaults to NULL.}

\item{prob_levels}{the probability of being assigned to a particular category for categorical variables (provides some ability to control proportions of individuals in each category).}
}
\value{
A nested list. The first element of the list contains the dataframe of individual data. Each subsequent element contains a list of information about each explanatory variable added.
}
\description{
Generates trait data for individuals in the population. Adds one variable at a time sequentially.
}
\details{
Generation of trait data (e.g. sex, body size etc.) for individuals in the population. Each time the function is run it adds one variable at a time sequentially. to the existing object.
}
