% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_info_merge.R
\name{indiv_info_merge}
\alias{indiv_info_merge}
\title{indiv_info_merge}
\usage{
indiv_info_merge(indiv_infoA, indiv_infoB)
}
\arguments{
\item{indiv_infoA}{first dataframe}

\item{indiv_infoB}{second dataframe}
}
\value{
merged dataframe
}
\description{
Merges two individual_info dataframes
}
