% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generation_seq.R
\name{interaction_generation_seq}
\alias{interaction_generation_seq}
\title{interaction_generation_seq}
\usage{
interaction_generation_seq(
  pop_mat,
  indiv_data,
  ne = 1000,
  mgs = 2,
  style_gs = c("rep", "pois"),
  par = 20,
  float = 1e-05
)
}
\arguments{
\item{pop_mat}{adjacency matrix of underlying social network}

\item{indiv_data}{indiv_data dataframe including information on IDs, group membership and spatial locations}

\item{ne}{number of groups}

\item{mgs}{mean group size}

\item{style_gs}{group size distribution. Currently either identical (style_gs=”rep”) or Poisson (style_gs=”pois”).}

\item{par}{used in jpc to adjust social network connection strengths to probabilities of being sampled in the same group. Defaults to 20}

\item{float}{used in jpc to provide a minimum probability of being sampled in the same group to allow all groups to be filled. Defaults to 0.0001}
}
\value{
The GBI generated by the function
}
\description{
Generates a sequential series of groups
}
\details{
Generates a sequential series of groups (c.f. interaction_generation_simul), whereby each group occurs at a different timestep so can contains overlapping sets of individuals.
}
