% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generation_simul.R
\name{interaction_generation_simul}
\alias{interaction_generation_simul}
\title{interaction_generation_simul}
\usage{
interaction_generation_simul(
  indiv_data,
  pop_mat,
  mean_group_size = 2,
  n_ts = 20,
  float = 1e-09,
  pm = 50,
  pow = 4
)
}
\arguments{
\item{indiv_data}{indiv_data dataframe containing information on ID, group membership and spatial locations}

\item{pop_mat}{adjacency matrix for underlying social network structure}

\item{mean_group_size}{mean group size. Defaults to 2.}

\item{n_ts}{number of time-steps to simulate grouping events for. Defaults to 20.}

\item{float}{used in jpc to provide a minimum probability of being sampled in the same group to allow all groups to be filled. Defaults to 0.000000001.}

\item{pm}{used in jpc to adjust social network connection strengths to probabilities of being sampled in the same group. Defaults to 50.}

\item{pow}{further adjustment of probabilities of being sampled in the same group based on social network connections. Defaults to 4.}
}
\value{
A list with two elements: A) the GBI generated by the function; B) a vector indicating which timestep each group occurred in.
}
\description{
Generates a series of simultaneously observed groups
}
\details{
Generates groups that exist contemporaneously (c.f. interaction_generation_seq), whereby individuals are only present in a single group. Note that the functionality for having close control over the group size distribution is currently limited and could be improved. Group sizes could instead be generated directly from a series of breaks provided if more control is desired.
}
