% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpc.R
\name{jpc}
\alias{jpc}
\title{jpc}
\usage{
jpc(group, par, float)
}
\arguments{
\item{group}{column of matrix (see use)}

\item{par}{used to adjust social network connection strengths to probabilities of being sampled in the same group.}

\item{float}{used to provide a minimum probability of being sampled in the same group to allow all groups to be filled.}
}
\value{
the join probabilities for candidates to join an interaction
}
\description{
Internal function used in interaction generation
}
