% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_checker_seq.R
\name{network_checker_seq}
\alias{network_checker_seq}
\title{network_checker_seq}
\usage{
network_checker_seq(gbi, pop_net)
}
\arguments{
\item{gbi}{group-by-individual matrix generated from the interaction generation function}

\item{pop_net}{underlying network used to generate GBI}
}
\value{
1) a plot of the network generated from interaction data; 2) plots of number of observations in the GBI vs. metrics in the underlying network; 3) an estimate of the correlation between the two networks; 4) plots of metric correlations between the two networks
}
\description{
Used to check how well sequential interaction data reflect underlying network
}
\details{
This function can be used to test various properties of the social network generated from the interaction dataset to see how closely it is related to the original underlying network.
}
