% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation_basic.R
\name{network_generation_basic}
\alias{network_generation_basic}
\title{network_generation_basic}
\usage{
network_generation_basic(
  indiv_data,
  dist_mat,
  p_ig = 0.2,
  w_ig1 = 1.5,
  w_ig2 = 5,
  p_og = 0.2,
  w_og1 = 1,
  w_og2 = 5,
  d_effp = 4,
  d_effw = 4,
  plot = TRUE
)
}
\arguments{
\item{indiv_data}{the indiv_data dataframe providing information on social group ID, spatial location etc.}

\item{dist_mat}{the distance matrix between all individuals in the population}

\item{p_ig}{probability of a within-group edge}

\item{w_ig1}{first parameter of the beta distribution controlling within-group edge weights}

\item{w_ig2}{second parameter of the beta distribution controlling within-group edge weights}

\item{p_og}{(intercept) probability of a between-group edge}

\item{w_og1}{first parameter of the beta distribution controlling between-group edge weights (intercept)}

\item{w_og2}{second parameter of the beta distribution controlling between-group edge weights (intercept)}

\item{d_effp}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{d_effw}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{plot}{Whether to plot the network generated. Defaults to TRUE.}
}
\value{
List of two parts: A) adjacency matrix and B) igraph network object
}
\description{
Generates underlying social network structure (weighted edges) based on spatial locations and group membership.
}
\details{
Probabilities of edges within and between groups (binomial). Weights of edges within and between groups (Beta, 2 parameter). Edge probabilities and weights depend on distance between groups. Note that for this function (as opposed to network_generation_covariates) you parameterise the beta distribution for edge weights directly.
}
