% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation_covariates.R
\name{network_generation_covariates}
\alias{network_generation_covariates}
\title{network_generation_covariates}
\usage{
network_generation_covariates(
  indiv_data,
  dist_mat,
  indiv_info,
  p_ig = 0.2,
  wi_m = 0.25,
  wi_v = 0.025,
  p_og = 0.2,
  wo_m = 0.25,
  wo_v = 0.025,
  d_effp = 4,
  d_effw = 4,
  covs = 2,
  effs,
  plot = TRUE
)
}
\arguments{
\item{indiv_data}{the indiv_data dataframe providing information on social group ID, spatial location etc.}

\item{dist_mat}{the distance matrix between all individuals in the population}

\item{indiv_info}{the indiv_info dataframe containing information on individual traits}

\item{p_ig}{probability of a within-group edge}

\item{wi_m}{mean within-group edge weight to generate a beta distribution of potential values}

\item{wi_v}{variance in within-group edge weight to generate a beta distribution of potential values}

\item{p_og}{probability of a between-group edge}

\item{wo_m}{mean between-group edge weight to generate a beta distribution of potential values}

\item{wo_v}{variance in within-group edge weight to generate a beta distribution of potential values}

\item{d_effp}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{d_effw}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{covs}{the columns of indiv_info used as covariates when generating the network}

\item{effs}{list of effect sizes (each element as vector for covariate or matrix for factor) containing effect of each trait (given by covs=) on each of the 6 parameters (p_ig, wi_m, wi_v, p_og, wo_m, wo_v, d_effp, d_effw)}

\item{plot}{Whether to plot the network generated. Defaults to TRUE.}
}
\value{
List of two parts: A) adjacency matrix and B) igraph network object
}
\description{
Generates underlying social network structure (weighted edges) based on spatial locations, group membership and individual traits.
}
\details{
More complex social network generation based on individual characteristics. Probabilities of edges within and between groups (binomial). Weights of edges within and between groups (Beta, 2 parameter). Edge probabilities and weights depend on distance between groups. Additions from network_generation_basic() are that you can use covariates from indiv_info to generate differences in network behaviour. Note that this function is parameterised using the mean and variance of the beta distribution so is more intuitive but perhaps a little less flexible. There is currently no functionality for interactions between individual traits but this could be added.There is currently no functionality for more complex network effects (e.g. transitivity, homophily etc.) but this is planned.
}
