% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_rewire_covariates.R
\name{network_rewire_covariates}
\alias{network_rewire_covariates}
\title{network_rewire_covariates}
\usage{
network_rewire_covariates(
  network,
  indiv_data,
  prev_data,
  dist_mat,
  indiv_info,
  p_ig = 0.2,
  wi_m = 0.25,
  wi_v = 0.025,
  p_og = 0.2,
  wo_m = 0.25,
  wo_v = 0.025,
  d_effp = 4,
  d_effw = 4,
  covs = 2,
  effs,
  p_wr_i = 0.2,
  p_wr_e = 0.5,
  plot = TRUE
)
}
\arguments{
\item{network}{the network to be rewired}

\item{indiv_data}{the current iteration (i.e. after survival/recruitment simulated) indiv_data dataframe providing information on social group ID, spatial location etc.}

\item{prev_data}{the previous iteration (i.e. before survival/recruitment simulated) indiv_data dataframe providing information on social group ID, spatial location etc.}

\item{dist_mat}{the distance matrix between all individuals in the population}

\item{indiv_info}{the indiv_info dataframe containing information on individual traits}

\item{p_ig}{probability of a within-group edge}

\item{wi_m}{mean within-group edge weight to generate a beta distribution of potential values}

\item{wi_v}{variance in within-group edge weight to generate a beta distribution of potential values}

\item{p_og}{probability of a between-group edge}

\item{wo_m}{mean between-group edge weight to generate a beta distribution of potential values}

\item{wo_v}{variance in within-group edge weight to generate a beta distribution of potential values}

\item{d_effp}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{d_effw}{distance effect of the probability of between-group edges. Implemented as (1/d_eff)^distance.}

\item{covs}{the columns of indiv_info used as covariates when generating the network}

\item{effs}{list of effect sizes (each element as vector for covariate or matrix for factor) containing effect of each trait (given by covs=) on each of the 6 parameters (p_ig, wi_m, wi_v, p_og, wo_m, wo_v, d_effp, d_effw)}

\item{p_wr_i}{probability of an existing individual have its existing social relationships in the network re-wired}

\item{p_wr_e}{probability that each edge is rewired for a selected individual.}

\item{plot}{Whether to plot the network generated. Defaults to TRUE.}
}
\value{
List of two parts: A) adjacency matrix and B) igraph network object
}
\description{
A function to add new individuals to and rewire a social network at each demographic timestep (or more frequently if desired).
}
\details{
Function to add new individuals to and rewire a social network. Uses the same structure as network_generation_covariates. The same caveats apply to this function as the network_generation_covariates function.
}
