% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_generation_basic.R
\name{population_generation_basic}
\alias{population_generation_basic}
\title{population_generation_basic}
\usage{
population_generation_basic(n, ng, plot = TRUE)
}
\arguments{
\item{n}{the total number of individuals in the population}

\item{ng}{the number of social groups the individuals should be divided into}

\item{plot}{Whether to plot spatial locations of groups. Defaults to TRUE.}
}
\value{
A list of two parts. A) A dataframe providing information on individual IDs, spatial coordinates and social group IDs. B) A distance matrix providing the spatial distances between all individuals in the population.
}
\description{
Generates basic data for a population clustered in social groups distributed uniformly in 2D space.
}
\details{
Function for basic population generation.Integrates information on number of individuals, groups and group locations. When group sizes are greater than 1 individuals are sampled into groups randomly, so not all groups will necessarily be identical in size. Currently the spatial grid remains the same size regardless of population size.
}
