% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestep_demographics.R
\name{timestep_demographics}
\alias{timestep_demographics}
\title{timestep_demographics}
\usage{
timestep_demographics(
  indiv_data,
  full_indiv_data,
  recruitment = TRUE,
  mps = 0.95,
  lvps = 0.5
)
}
\arguments{
\item{indiv_data}{the current indiv_data dataframe (including data on individual ID, social group, spatial location and survival)}

\item{full_indiv_data}{the current full_indiv_data dataframe containing the information for all individuals present during the study}

\item{recruitment}{= indicates whether recruitment into the population is possible. Defaults to TRUE}

\item{mps}{the mean population survival. Defaults to NULL at which point it is calculated from indiv_data$survival.}

\item{lvps}{the variation in survival probability used to generate initial values in the function. Relatively unimportant for most uses of the code where this information is replaced subsequently. Defaults to 0.5.}
}
\value{
a list of three parts. 1)  The new iteration of indiv_data; 2) the updated version of full_indiv_data; 3) a new distance matrix for the updated population.
}
\description{
Survival and recruitment of individuals at the end of each demographic timestep.
}
\details{
Survival is stochastic based on each individual’s survival probability. The recruitment rate is calculated from population mean survival probability and is then implemented stochastically via a draw from a Poisson distribution. Note that for recruitment into social groups: a) when there are no social groups (each individual is its own group) then individuals are recruited into empty group locations to replace dead individuals where possible and then if not new locations are sampled; b) When there are social groups then individuals are sampled into social groups weighted by the inverse of group size (individuals more likely to be recruited into smaller groups).
}
