% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_checker_simul.R
\name{network_checker_simul}
\alias{network_checker_simul}
\title{network_checker_simul}
\usage{
network_checker_simul(gbi, pop_net)
}
\arguments{
\item{gbi}{group-by-individual matrix generated from the interaction generation function}

\item{pop_net}{underlying network used to generate GBI}
}
\value{
1) a plot of the network generated from interaction data; 2) an estimate of the correlation between the two networks; 3) plots of metric correlations between the two networks
}
\description{
Used to check how well simultaneous interaction data reflect underlying network
}
\details{
This function can be used to test various properties of the social network generated from the interaction dataset to see how closely it is related to the original underlying network.
}
